refactor: trace models and create standalone execution trace function
features:

- Refactored execution tracing into a separate function
- Added standalone `trace()` function usage without requiring engine instantiation.

internal :

- Added  tools for index file generation

⚠️ Deprecation Notice:

- `TraceableExecution` is deprecated and will be removed in a future release. Use `TraceableEngine` instead.

- `NodeExecutionTrace` is deprecated,  Use `ExecutionTrace` instead.

- `NodeExecutionTraceExtractor` is deprecated. Use `ExecutionTraceExtractor` instead.

- `isNodeExecutionTrace` is deprecated. Use `isExecutionTrace` instead.
