+++ examples/greeting.json
@@ -1,7 +1,7 @@
 [
   {
     "data": {
-      "id": "generateGreeting_1739227865148_820f108e-62b9-4f3d-a768-2b5019bffd0c",
+      "id": "generateGreeting_1739477824253_2f5d9461-e3ab-433b-a19b-2e18642718fb",
       "label": "1 - generateGreeting",
       "inputs": [
         {
@@ -21,9 +21,9 @@
       "narratives": [
         "here is tracing narrative for greeting John Doe"
       ],
-      "startTime": "2025-02-10T22:51:05.148Z",
+      "startTime": "2025-02-13T20:17:04.254Z",
       "abstract": false,
-      "createTime": "2025-02-10T22:51:05.148Z"
+      "createTime": "2025-02-13T20:17:04.254Z"
     },
     "group": "nodes"
   }
