+++ src/common/awaited.ts
@@ -1,0 +1,1 @@
+export type Awaited<T> = T extends PromiseLike<infer U> ? U : T;
