+++ src/index.ts
@@ -1,5 +1,14 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
 export * from './engine/executionEngine';
 export * from './engine/executionEngineDecorators';
-export * from './trace/trace.model';
-export * from './trace/traceableExecution';
+export * from './engine/traceableEngine';
 export * from './timer/executionTimer';
+export * from './trace/trace';
+export * from './trace/traceableExecution';
+export * from './trace/models/engineNodeTrace.model';
+export * from './trace/models/engineTrace.model';
+export * from './trace/models/engineTraceOptions.model';
+export * from './trace/models/executionTrace.model';
